# Backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RapidReachNET, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/admin/addEmployee": {
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeeCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/updateEmployee/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeeUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/getAllUsers": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/getAllEmployees": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/getEmployeeById/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/addBranch": {
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Branch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/getAllBranches": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/admin/getBranchById/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/addCourier/{customerId}": {
      "post": {
        "tags": [
          "Courier"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CourierRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getCourierById/{id}": {
      "get": {
        "tags": [
          "Courier"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getCouriersByUserId/{userId}": {
      "get": {
        "tags": [
          "Courier"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getCouriersByEmployee/{userId}": {
      "get": {
        "tags": [
          "Courier"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/addFeedback": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "comment",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        }
      }
    },
    "/customer/getAllFeedback": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FeedbackDTO"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Lgin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/addPayment": {
      "post": {
        "tags": [
          "Payment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PaymentRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/GetAllPayments": {
      "get": {
        "tags": [
          "Payment"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/GetPaymentsByUserId/{userId}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getTrackingStatusByCourierId/{courierId}": {
      "get": {
        "tags": [
          "Tracking"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courierId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/updateStatus/{trackingId}": {
      "put": {
        "tags": [
          "Tracking"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trackingId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/registerUser": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/updateUser/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getAllUsers": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/customer/getUserById/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Branch": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "branchName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "Courier": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "dropLocation": {
          "type": "string"
        },
        "parcelDescription": {
          "type": "string"
        },
        "parcelWeight": {
          "format": "double",
          "type": "number"
        },
        "pickUpLocation": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "branchId": {
          "format": "int64",
          "type": "integer"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "branch": {
          "$ref": "#/definitions/Branch"
        },
        "payments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Payment"
          }
        },
        "trackings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tracking"
          }
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "CourierRequestDTO": {
      "type": "object",
      "properties": {
        "dropLocation": {
          "type": "string"
        },
        "parcelDescription": {
          "type": "string"
        },
        "parcelWeight": {
          "format": "double",
          "type": "number"
        },
        "pickUpLocation": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "branchId": {
          "format": "int64",
          "type": "integer"
        }
      }
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "EmployeeCreateDTO": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "contact": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "pincode": {
          "type": "string"
        },
        "branchId": {
          "format": "int64",
          "type": "integer"
        }
      }
    },
    "EmployeeUpdateDTO": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "contact": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "pincode": {
          "type": "string"
        },
        "branchId": {
          "format": "int64",
          "type": "integer"
        }
      }
    },
    "Feedback": {
      "type": "object",
      "properties": {
        "feedbackId": {
          "format": "int64",
          "type": "integer"
        },
        "comment": {
          "type": "string"
        },
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "FeedbackDTO": {
      "type": "object",
      "properties": {
        "feedbackId": {
          "format": "int64",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "LoginDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Payment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "paymentDate": {
          "$ref": "#/definitions/DateOnly"
        },
        "courierId": {
          "format": "int64",
          "type": "integer"
        },
        "courier": {
          "$ref": "#/definitions/Courier"
        }
      }
    },
    "PaymentRequestDTO": {
      "type": "object",
      "properties": {
        "courierId": {
          "format": "int64",
          "type": "integer"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "paymentDate": {
          "$ref": "#/definitions/DateOnly"
        }
      }
    },
    "Tracking": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "courierDate": {
          "$ref": "#/definitions/DateOnly"
        },
        "trackingStatus": {
          "type": "string"
        },
        "courierId": {
          "format": "int64",
          "type": "integer"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int64",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "contact": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "pincode": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "branchId": {
          "format": "int64",
          "type": "integer"
        },
        "couriers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Courier"
          }
        },
        "feedbacks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Feedback"
          }
        }
      }
    }
  }
}